Extract class WindowMenu from class MainMenu.
