+++ Dev/Editor/Effekseer/Effekseer.csproj
@@ -145,6 +145,7 @@
     <Compile Include="GUI\Menu\MainMenu.cs" />
     <Compile Include="GUI\Menu\Menu.cs" />
     <Compile Include="GUI\Menu\MenuItem.cs" />
+    <Compile Include="GUI\Menu\WindowMenu.cs" />
     <Compile Include="GUI\Network.cs" />
     <Compile Include="GUI\ThumbnailManager.cs" />
     <Compile Include="GUI\Viewer.cs" />
